+++ sdks/java/io/kafka/src/main/java/org/apache/beam/sdk/io/kafka/KafkaRecordCoder.java
@@ -22,7 +22,6 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.List;
 import org.apache.beam.sdk.coders.Coder;
-import org.apache.beam.sdk.coders.CoderException;
 import org.apache.beam.sdk.coders.KvCoder;
 import org.apache.beam.sdk.coders.StringUtf8Coder;
 import org.apache.beam.sdk.coders.StructuredCoder;
@@ -50,37 +49,22 @@ public class KafkaRecordCoder<K, V> extends StructuredCoder<KafkaRecord<K, V>> {
   }
 
   @Override
-  public void encode(KafkaRecord<K, V> value, OutputStream outStream)
-      throws CoderException, IOException {
-    encode(value, outStream, Context.NESTED);
+  public void encode(KafkaRecord<K, V> value, OutputStream outStream) throws IOException {
+    stringCoder.encode(value.getTopic(), outStream);
+    intCoder.encode(value.getPartition(), outStream);
+    longCoder.encode(value.getOffset(), outStream);
+    longCoder.encode(value.getTimestamp(), outStream);
+    kvCoder.encode(value.getKV(), outStream);
   }
 
   @Override
-  public void encode(KafkaRecord<K, V> value, OutputStream outStream, Context context)
-                         throws CoderException, IOException {
-    Context nested = context.nested();
-    stringCoder.encode(value.getTopic(), outStream, nested);
-    intCoder.encode(value.getPartition(), outStream, nested);
-    longCoder.encode(value.getOffset(), outStream, nested);
-    longCoder.encode(value.getTimestamp(), outStream, nested);
-    kvCoder.encode(value.getKV(), outStream, context);
-  }
-
-  @Override
-  public KafkaRecord<K, V> decode(InputStream inStream) throws CoderException, IOException {
-    return decode(inStream, Context.NESTED);
-  }
-
-  @Override
-  public KafkaRecord<K, V> decode(InputStream inStream, Context context)
-                                      throws CoderException, IOException {
-    Context nested = context.nested();
+  public KafkaRecord<K, V> decode(InputStream inStream) throws IOException {
     return new KafkaRecord<>(
-        stringCoder.decode(inStream, nested),
-        intCoder.decode(inStream, nested),
-        longCoder.decode(inStream, nested),
-        longCoder.decode(inStream, nested),
-        kvCoder.decode(inStream, context));
+        stringCoder.decode(inStream),
+        intCoder.decode(inStream),
+        longCoder.decode(inStream),
+        longCoder.decode(inStream),
+        kvCoder.decode(inStream));
   }
 
   @Override
